Bullet Animation & Targeting (#22)
## Description
Improved the animation of the bullets by adding a less abrupt ending to
the bullet lifetime and including a manual ending to the animation on
the target being killed. Additionally I decided to drive the animation
by using their velocity to calculate a duration on start to create
variation. This can lead to some issues with moving targets creating
bullets at a later time to be completed first as they are closer than
older bullets.

Changed the attack component to not require the animation component to
work, and to avoid constantly selecting a target, now target selection
only occurs when the current target becomes invalid.

Fixed some animation bugs due to not having unique resources.

> [!WARNING]
> Bullet may be renamed to projectile in next PR, as well as other
refactoring.

## Summary of Changes
#### Animation
- Created `NotifiedAndAnimationNode3D` extending `NotifiedNode3D` to
support creating and firing off an animation on `AnimationPlayer` when
lifetime is coming to an end
- Updated bullet scene root to a `NotifiedAndAnimationNode3D` to animate
bullet mesh when the bullet is set for destruction
- Removed `travel_speed` from `BulletAnimation` as progress is now
driven by velocity instead of duration.
- Changed `AttackAnimationComponent` to generate an unique animation
resource to emit unique signals and fix visual bug.
- Updated `AttackAnimationComponent` and `AttackAnimation` to include
the option of ending an animation and emitting it.
- Updated `AttackAnimationComponent` and `AttackAnimation` to include
the option of forcefully stopping an animation.
- Added calculating and the duration based on velocity in
`AttackAnimationData`
- Removed unused variable in `ParticlesComponent`

#### Gameplay
- Updated `AttackComponent` to select and cache target until selected
target becomes invalid by dying
- Replaced `attack_duration` in `AttackComponent` with `attack_velocity`
to be used on animations when relevant
- Added a way to check if given damage value would be lethal to
`HealthComponent`
- Added comments to `HealthComponent`

### Bug Fixing
- Bullets no longer speed up after more than one tower is placed